requirements.txt
boto3==1.35.39
django-storages==1.14.4

# Django
# ------------------------------------------------------------------------------
django==5.0.9 
django-environ==0.11.2
# Django REST Framework
djangorestframework==3.15.2 
django-cors-headers==4.4.0  
drf-yasg==1.21.7
